Bump version to v9.1.0 plus fix a top_n query

+++ src/bot/user_event.rs
@@ -324,7 +324,7 @@ pub async fn process_user_message<S: SlackSender>(
                                     OrdQuery::Asc,
                                     KarmaTyp::Total,
                                     ("highest karma", "lowest karma"),
-                                    3u16,
+                                    3u32,
                                 )
                                 .await;
                             }
@@ -338,7 +338,7 @@ pub async fn process_user_message<S: SlackSender>(
                                     OrdQuery::Asc,
                                     KarmaTyp::Total,
                                     ("most positive", "most negative"),
-                                    3u16,
+                                    3u32,
                                 )
                                 .await;
                             }
@@ -352,7 +352,7 @@ pub async fn process_user_message<S: SlackSender>(
                                     OrdQuery::Desc,
                                     KarmaTyp::Side,
                                     ("most sidevotes received", "most sidevotes given"),
-                                    3u16,
+                                    3u32,
                                 )
                                 .await;
                             }
@@ -375,7 +375,7 @@ pub async fn process_user_message<S: SlackSender>(
                 Ok(Command(c @ ("topkarma" | "topgivers" | "topsidevotes"), arg)) => {
                     if arg.len() == 1 {
                         // Parse the argument
-                        let limit = u16::from_str(arg.first().unwrap_or(&"1"));
+                        let limit = u32::from_str(arg.first().unwrap_or(&"1"));
 
                         match (c, limit) {
                             ("topkarma", Ok(lim @ 1..=25)) => {
